+++ src/core/error_handling.c
@@ -24,9 +24,26 @@ PyodideErr_SetJsError(JsRef err)
   PyErr_SetObject((PyObject*)(py_err->ob_type), py_err);
 }
 
+PyObject* internal_error;
+PyObject* conversion_error;
+
 int
-error_handling_init()
+error_handling_init(PyObject* core_module)
 {
+  bool success = false;
+  internal_error = PyErr_NewException("pyodide.InternalError", NULL, NULL);
+  FAIL_IF_NULL(internal_error);
+
+  conversion_error = PyErr_NewExceptionWithDoc(
+    "pyodide.ConversionError",
+    PyDoc_STR("Raised when conversion between Javascript and Python fails."),
+    NULL,
+    NULL);
+  FAIL_IF_NULL(conversion_error);
+  // ConversionError is public
+  FAIL_IF_MINUS_ONE(
+    PyObject_SetAttrString(core_module, "ConversionError", conversion_error));
+
   EM_ASM({
     Module.handle_js_error = function(e)
     {
@@ -35,5 +52,8 @@ error_handling_init()
       Module.hiwire.decref(err);
     };
   });
-  return 0;
+
+  success = true;
+finally:
+  return success ? 0 : -1;
 }